Python/requirements.txt
async-timeout==4.0.3
autopep8==1.5.6
backcall==0.2.0
bleak==0.21.1
bleak-winrt==1.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.4
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.4.5
matplotlib==3.7.2
mido==1.3.0
mllib==1.0.0a2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.1
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.16
pycodestyle==2.7.0
Pygments==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-rtmidi==1.5.5
pytz==2023.3.post1
pywin32==300
pyzmq==22.0.3
requests==2.31.0
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.5
zipp==3.16.2
